Split and rename JavaNetHttp and ApacheHttp tests for consistency
